LayoutTests/fast/dom/prototype-property-not-enumerable.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("The 'prototype' property on interface objects should have attributes { [[Writable]]: false, [[Enumerable]]: false, [[Configurable]]: false }");

debug("* Event");
var descriptor = Object.getOwnPropertyDescriptor(Event, "prototype");
shouldBeFalse("descriptor.writable");
shouldBeFalse("descriptor.enumerable");
shouldBeFalse("descriptor.configurable");

debug("* Element");
var descriptor = Object.getOwnPropertyDescriptor(Element, "prototype");
shouldBeFalse("descriptor.writable");
shouldBeFalse("descriptor.enumerable");
shouldBeFalse("descriptor.configurable");

debug("* Document");
var descriptor = Object.getOwnPropertyDescriptor(Document, "prototype");
shouldBeFalse("descriptor.writable");
shouldBeFalse("descriptor.enumerable");
shouldBeFalse("descriptor.configurable");

debug("* Window");
var descriptor = Object.getOwnPropertyDescriptor(Window, "prototype");
shouldBeFalse("descriptor.writable");
shouldBeFalse("descriptor.enumerable");
shouldBeFalse("descriptor.configurable");

debug("* HTMLOptionElement");
var descriptor = Object.getOwnPropertyDescriptor(HTMLOptionElement, "prototype");
shouldBeFalse("descriptor.writable");
shouldBeFalse("descriptor.enumerable");
shouldBeFalse("descriptor.configurable");

debug("* Option");
var descriptor = Object.getOwnPropertyDescriptor(Option, "prototype");
shouldBeFalse("descriptor.writable");
shouldBeFalse("descriptor.enumerable");
shouldBeFalse("descriptor.configurable");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
